slogging through refactor
