+++ app/build.gradle
@@ -25,7 +25,6 @@ dependencies {
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    implementation project(":base")
     implementation project(":cardscan")
     implementation 'com.stripe:stripe-android:8.7.0'
 }
