+++ cardscan/build.gradle
@@ -28,6 +28,7 @@ dependencies {
     implementation 'com.android.support:support-annotations:28.0.0'
     implementation "com.android.support:appcompat-v7:28.0.0"
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    //implementation 'com.getbouncer:base:1.0.4012'
     implementation project(":base")
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
@@ -50,7 +51,7 @@ ext {
     siteUrl = 'https://github.com/getbouncer/cardscan-android'
     gitUrl = 'https://github.com/getbouncer/cardscan-android.git'
 
-    libraryVersion = '1.0.4011'
+    libraryVersion = '1.0.4012'
 
     developerId = 'kingst'
     developerName = 'Sam King'
